### 2021-04-09 #3164 @jvanhoefer
Oh, I apologize, it is in this repository

https://github.com/yaml2sbml-dev/paper

(I gave the link to the toolbox and not to the paper...)
