### 2021-04-10 #3164 @danielskatz
For JOSS, we need the paper and the software to be in the same repository, though if you need, the paper can be in a branch that doesn't gets merged. For most submissions, it's just in a paper or joss directory in the software repo.
