+++ doc/JOSS_paper/paper.bib
@@ -1,0 +1,302 @@
+@article{HuckaFin2003,
+	Author = {Hucka, M. and Finney, A. and Sauro, H. M. and Bolouri, H. and Doyle, J. C. and Kitano, H. and Arkin, A. P. and Bornstein, B. J. and Bray, D. and Cornish-Bowden, A. and Cuellar, A. A. and Dronov, S. and Gilles, E. D. and Ginkel, M. and Gor, V. and Goryanin, I. I. and Hedley, W. J. and Hodgman, T. C. and Hofmeyr, J.-H. and Hunter, P. J. and Juty, N. S. and Kasberger, J. L. and Kremling, A. and Kummer, U. and {Le Nov\`{e}re}, N. and Loew, L. M. and Lucio, D. and Mendes, P. and Minch, E. and Mjolsness, E. D. and Nakayama, Y. and Nelson, M. R. and Nielsen, P. F. and Sakurada, T. and Schaff, J. C. and Shapiro, B. E. and Shimizu, T. S. and Spence, H. D. and Stelling, J. and Takahashi, K. and Tomita, M. and Wagner, J. and Wang, J.},
+	Date-Added = {2016-08-17 16:49:31 +0000},
+	Date-Modified = {2018-04-13 20:51:14 +0000},
+	Doi = {10.1093/bioinformatics/btg015},
+	Journal = {Bioinformatics},
+	Journal-Full = {Bioinformatics},
+	Keywords = {SBML},
+	Months = {March},
+	Number = {4},
+	Pages = {524--531},
+	Title = {The systems biology markup language {(SBML):} {A} medium for representation and exchange of biochemical network models},
+	Volume = {19},
+	Year = {2003},
+	Bdsk-File-1 = {YnBsaXN0MDDSAQIDBFxyZWxhdGl2ZVBhdGhZYWxpYXNEYXRhXxAYUGFwZXJzL0h1Y2thRmluMjAwM2EucGRmTxEBiAAAAAABiAACAAAMTWFjaW50b3NoIEhEAAAAAAAAAAAAAAAAAAAAAAAAAEJEAAH/////EUh1Y2thRmluMjAwM2EucGRmAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAP////8AAAAAAAAAAAAAAAAAAQADAAAKIGN1AAAAAAAAAAAAAAAAAAZQYXBlcnMAAgBGLzpVc2VyczpqYW4uaGFzZW5hdWVyOnNjaWVibzpNTFM6QmlibGlvZ3JhcGh5OlBhcGVyczpIdWNrYUZpbjIwMDNhLnBkZgAOACQAEQBIAHUAYwBrAGEARgBpAG4AMgAwADAAMwBhAC4AcABkAGYADwAaAAwATQBhAGMAaQBuAHQAbwBzAGgAIABIAEQAEgBEVXNlcnMvamFuLmhhc2VuYXVlci9zY2llYm8vTUxTL0JpYmxpb2dyYXBoeS9QYXBlcnMvSHVja2FGaW4yMDAzYS5wZGYAEwABLwAAFQACABT//wAAAAgADQAaACQAPwAAAAAAAAIBAAAAAAAAAAUAAAAAAAAAAAAAAAAAAAHL},
+	Bdsk-Url-1 = {http://dx.doi.org/10.1093/bioinformatics/btg015}}
+	
+@article{HoopsSah2006,
+	Author = {Hoops, S. and Sahle, S. and Gauges, R. and Lee, C. and Pahle, J. and Simus, N. and Singhal, M. and Xu, L. and Mendes, P. and Kummer, U.},
[…]
