+++ doc/JOSS_paper/paper.md
@@ -1,0 +1,128 @@
+---
+title: "yaml2sbml: Human-readable and -writable specification of ODE models and their conversion to SBML"
+tags:
+  - Python
+  - SBML
+  - Ordinary Differential Equation
+  - YAML
+authors:
+  - name: Jakob Vanhoefer
+    orcid: 0000-0002-3451-1701
+    affiliation: 1
+  - name: Marta R. A. Matos
+    orcid: 0000-0003-4288-1005
+    affiliation: 2
+  - name: Dilan Pathirana
+    orcid: 0000-0001-7000-2659
+    affiliation: 1
+  - name: Yannik Schälte
+    orcid: 0000-0003-1293-820X
+    affiliation: "3, 4"
+  - name: Jan Hasenauer
+    orcid: 0000-0002-4935-3312
+    affiliation: "1, 3, 4"
+affiliations:
+    - name: Faculty of Mathematics and Natural Sciences, University of Bonn, Bonn, Germany
+      index: 1
+    - name: The Novo Nordisk Foundation Center for Biosustainability, Technical University of Denmark, DK-2800 Kgs. Lyngby, Denmark
+      index: 2
+    - name: Institute of Computational Biology, Helmholtz Zentrum München, Neuherberg, Germany
+      index: 3
+    - name: Department of Mathematics, Technical University Munich, Garching, Germany
+      index: 4
+date: 05 March 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+Ordinary differential equations (ODE)  models are used throughout natural sciences to describe dynamic processes. In  systems biology, ODEs are mostly stored and exchanged using the Systems Biology Markup Language (SBML),  a widely adopted community standard based on XML. The Parameter Estimation table (PEtab) format extends SBML to parameter estimation problems. A large number of software tools support simulation of SBML models and parameter estimation for PEtab problems. Specifying ODE models in SBML and parameter estimation problems in PEtab provides access to these tools. However, SBML is considered to be neither human-readable nor human-writable. An easy-to-use approach to construct the SBML/PEtab models tailored to ODE models will facilitate model generation.
+
+In this contribution, we present `yaml2sbml`, a Python tool for converting ODE models specified in an easy-to-read and -write YAML file into SBML/PEtab.  `yaml2sbml` comes with a format validator for the input YAML, a command-line interface (CLI) and a model editor to: 1) create an ODE model programmatically in Python that can then be saved as SBML, PEtab or YAML and 2) edit an ODE model previously encoded in YAML. Several examples illustrate the use of `yaml2sbml` on realistic problems.
+
+# Statement of need
+
+ODE models have applications in many scientific fields, including biology [@HodgkinHux1952; @Lotka1920],  epidemiology [@KermackMcK1927], physics [@VanderPol1927; @Lorenz1963], economics [@Shone2002] or more recent neural networks [@ChenRub2018].
+
+SBML is a widely adopted community standard for specifying biological reaction networks [@HuckaFin2003]. [sbml.org](http://sbml.org/SBML_Software_Guide/SBML_Software_Summary#cat_12) lists more than 100 software tools that accept SBML as their input format for dynamic model simulation, among them COPASI [@HoopsSah2006], d2d [@RaueSte2015], AMICI [@FroehlichWei2020] and dMod [@KaschekMad2019].
+
[…]
