### 2021-04-12 #3164 @jvanhoefer
Thank you for the info, that wasn't clear to me. The paper is now in

https://github.com/yaml2sbml-dev/yaml2sbml/tree/JOSS_review/doc/JOSS_paper

(i.e. in folder `doc/JOSS_paper` of the `JOSS_review` branch.)
