### 2021-04-23 #3215 @whedon
[REVIEW]: yaml2sbml: Human-readable and -writable specification of ODE models and their conversion to SBML
**Submitting author:** @jvanhoefer (<a href="http://orcid.org/0000-0002-3451-1701">Jakob Vanhoefer</a>)
**Repository:** <a href="https://github.com/yaml2sbml-dev/yaml2sbml" target ="_blank">https://github.com/yaml2sbml-dev/yaml2sbml</a>
**Version:** v0.2.3
