### 2021-05-01 yaml2sbml-dev/yaml2sbml#125 @SirSharpest
Reviewer comments for JOSS submisson
Hi,

I'll be making any comments, suggestions or issue I run into here, to keep the main review page less-cluttered. This is in relation to the main review page here: https://github.com/openjournals/joss-reviews/issues/3215

I'll be making comments as I come across them in the next few days (apologies for any delay, I am currently travelling).

Thanks for your submission, and I look forward to reviewing this submission.

Best,
Nathan
