### 2021-05-01 yaml2sbml-dev/yaml2sbml#125 @SirSharpest
I'd like to start by saying that the quality of this repo is excellent. Documentation is great, for example when I wanted to find information on running tests I found it to be perfectly explained.

Some minor thoughts:

- The MIT license names two authors and not all authors named on the paper. Is this intentional, if so, disregard this comment.
-  Your examples have some extra dependencies, I recognize that these are configurable in the setup file, but I could not find mention of the specific ```pip command```  in the docs to run to get example notebooks to run quickly.
- I ran into some issues with the amici library installing on my OSX machine https://amici.readthedocs.io/en/v0.11.12/python_installation.html - there are some additional dependencies, possibly worth flagging as other users may come across this when doing a clean installation?
    - I recognise this is not an issue of your software, but just may create a barrier for users who may quit trying to use the package after a wall of errors
