### 2021-05-02 yaml2sbml-dev/yaml2sbml#125 @jvanhoefer
Hey Nathan,

first of all: Thank you a lot for your positive and constructive feedback! I will address your comments in the branch `feature_JOSS_review_comments`.

Regarding AMICI: You are right, AMICI can have some issues during installation and it will be helpful to offer some guidance for installing AMICI already in the docs of yaml2sbml.

Did you get AMICI running in the end? (AMICI is also developed in our group, therefore if you have any questions regarding that: We are happy to help :) )
