+++ doc/release_notes.rst
@@ -8,7 +8,7 @@ Release Notes
 0.2.2 (2021-03-19)
 ------------------
 
-* Add checks in SBML conversion, e.g. to catch invalid identifiers and equations (# 118).
+* Add checks in SBML conversion, e.g. to catch invalid identifiers and equations (# 118, #123).
 * Add tests for MacOS and Windows (#115).
 * Fix pydocstyle and swig (#120).
 * Add a logo license. (#116)
