### 2021-05-04 yaml2sbml-dev/yaml2sbml#125 @SirSharpest
Yeah, I did, thanks.

As I said, it was very difficult to find any improvements which could be made. I'm confident that this is ready for publication.
