### 2021-05-04 #3215 @SirSharpest
@kellyrowland I've completed my review + checklist, this software package is an excellent example of open source software. I have given the authors some minor suggestions regarding potential issues which users on certain OS's may encounter. I am very happy to recommend acceptance.
