### 2021-05-13 yaml2sbml-dev/yaml2sbml#127 @marouenbg
Reviewer 2 comments for JOSS submission
yaml2sbml's implementation was executed flawlessly with a complete set of docs and several examples.
These are my comments in relation the review process https://github.com/openjournals/joss-reviews/issues/3215

# Comments on the paper:
These are a few suggestions for the authors about the main text.

- Currently 9 software toolboxes support PEtab as an input format, among them COPASI, d2d,
Would it be possible to complete the list?
- Due to the aforementioned tools, model simulation or parameter estimation has become a matter of a few lines of code or clicks
I suggest using Thanks to the aforementioned tools

- Hence ODE model definition is a relevant bottleneck,
I would suggest: However, ODE model definition is often a bottleneck

- Python-based model editor that allows one to generate
My suggestion: Python-based model editor that allows to generate

- Figure 1 gives an overview over of the typical
 Figure 1 gives an overview of the typical

 - Building the input format on YAML allows one to easily parse and validate the model
 Building the input format on YAML allows to parse and validate the model easily
