### 2021-05-13 yaml2sbml-dev/yaml2sbml#127 @marouenbg
# Comments on the tool

- specify the dependencies in the first tutorial - finite state projection ( amici package)
- the URL links in the beginning of Format features tutorial are not working
