+++ doc/JOSS_paper/paper.md
@@ -46,15 +46,15 @@ ODE models have applications in many scientific fields, including biology [@Hodg
 
 SBML is a widely adopted community standard for specifying biological reaction networks [@HuckaFin2003]. [sbml.org](http://sbml.org/SBML_Software_Guide/SBML_Software_Summary#cat_12) lists more than 100 software tools that accept SBML as their input format for dynamic model simulation, among them COPASI [@HoopsSah2006], d2d [@RaueSte2015], AMICI [@FroehlichWei2020] and dMod [@KaschekMad2019].
 
-Model parameters can be estimated from data by formulating a likelihood function. Therefore, the system states must be mapped to measured quantities by observable functions, and a measurement noise model must be specified. The PEtab format was recently introduced to complement SBML by tab-separated value files specifying observables, measurements, experimental conditions, and estimated parameters [@SchmiesterSch2021]. Currently 9 software toolboxes support PEtab as an input format, among them COPASI, d2d, dMod and AMICI/[pyPESTO](https://github.com/ICB-DCM/pyPESTO).
+Model parameters can be estimated from data by formulating a likelihood function. Therefore, the system states must be mapped to measured quantities by observable functions, and a measurement noise model must be specified. The PEtab format was recently introduced to complement SBML by tab-separated value files specifying observables, measurements, experimental conditions, and estimated parameters [@SchmiesterSch2021]. Currently 9 software toolboxes support PEtab as an input format, among them COPASI, d2d, dMod and AMICI/[pyPESTO](https://github.com/ICB-DCM/pyPESTO). The  [PEtab documentation](https://petab.readthedocs.io/en/latest/#petab-support-in-systems-biology-tools) gives a complete and up-to-date list of tools.
 
-Due to the aforementioned tools, model simulation or parameter estimation has become a matter of a few lines of code or clicks. Hence ODE model definition is a relevant bottleneck, since constructing an SBML model from scratch is often tedious. Therefore, various approaches to facilitate model construction from text-based input formats or in code have been presented, as `libsbml` [@BornsteinKea2008],  `SimpleSBML` [@CannistraMed2015], `MOCCASIN` [@GomezHuc2016], `Antimony` [@SmithBer2009] and `ScrumPy`  [@Poolman2006]. `MOCCASIN` translates MATLAB code into SBML. Other tools have a text-based input format that is centered around chemical reactions and not around ODEs directly (e.g. `ScrumPy`), or only offer a text-based (`Antimony`) or only a Python-based way of defining SBML models ( `libsbml`, `SimpleSBML`), but not both at the same time interchangeably. Neither of these tools offer PEtab support.
+Thanks to the aforementioned tools, model simulation or parameter estimation has become a matter of a few lines of code or clicks. However, ODE model definition is often a bottleneck, since constructing an SBML model from scratch is often tedious. Therefore, various approaches to facilitate model construction from text-based input formats or in code have been presented, as `libsbml` [@BornsteinKea2008],  `SimpleSBML` [@CannistraMed2015], `MOCCASIN` [@GomezHuc2016], `Antimony` [@SmithBer2009] and `ScrumPy`  [@Poolman2006]. `MOCCASIN` translates MATLAB code into SBML. Other tools have a text-based input format that is centered around chemical reactions and not around ODEs directly (e.g. `ScrumPy`), or only offer a text-based (`Antimony`) or only a Python-based way of defining SBML models ( `libsbml`, `SimpleSBML`), but not both at the same time interchangeably. Neither of these tools offer PEtab support.
 
[…]
