### 2021-05-24 yaml2sbml-dev/yaml2sbml#127 @jvanhoefer
Dear Marouen,

Thank you for your helpful comments and for reviewing our submission! I addressed your suggestions in two PRs for your comments regarding the tool (#128) and the text (#129).  I agree, that your reformulation will improve the manuscript! Regarding

> * Currently 9 software toolboxes support PEtab as an input format, among them COPASI, d2d,
Would it be possible to complete the list?

I added a link to the PEtab documentation, that gives a dynamically updated list with detailed overview over tools + support (as there are to my knowledge more tools who will have a PEtab support soon).
