### 2021-05-24 a69ea4a
Update doc/JOSS_paper/paper.md
Co-authored-by: Yannik Schälte <31767307+yannikschaelte@users.noreply.github.com>
