### 2021-05-24 a49f3b0
Update INSTALL.rst
Co-authored-by: Yannik Schälte <31767307+yannikschaelte@users.noreply.github.com>
