### 2021-05-24 6aa04f6
Update INSTALL.rst
Co-authored-by: Dilan Pathirana <59329744+dilpath@users.noreply.github.com>
