### 2021-05-24 c82485a
Adress comments regarding the software in the JOSS reviewing process (#128)
* comments reviewer 1

* delete links in Jupiter notebook

delete links, as they are not displayed correctly online on GitHub...

* Change relative to absolute GH link

The relative link to the file is not displayed correctly on GitHub -> change to absolute link on GitHub page...

* Update INSTALL.rst

Co-authored-by: Yannik Schälte <31767307+yannikschaelte@users.noreply.github.com>

* Update doc/examples/Format_Features/Format_Features.ipynb

Co-authored-by: Dilan Pathirana <59329744+dilpath@users.noreply.github.com>

[…]
