### 2021-05-25 #3215 @jvanhoefer
Is there anything else, that I need to do? I would add the Zenodo-Link/Badge together with a JOSS-link/badge in the next release directly after publication here (if there is no different procedure...) :)
