+++ doc/JOSS_paper/paper.bib
@@ -79,7 +79,7 @@
 	Year = {2006}}
 
 @article{SchmiesterSch2021,
-    author = {Schmiester, Leonard AND Sch{\"a}lte, Yannik AND Bergmann, Frank T. AND Camba, Tacio AND Dudkin, Erika AND Egert, Janine AND Fr{\"o}hlich, Fabian AND Fuhrmann, Lara AND Hauber, Adrian L. AND Kemmer, Svenja AND Lakrisenko, Polina AND Loos, Carolin AND Merkt, Simon AND M{\"u}ller, Wolfgang AND Pathirana, Dilan AND Raim{\'u}ndez, Elba AND Refisch, Lukas AND Rosenblatt, Marcus AND Stapor, Paul L. AND St{\"a}dter, Philipp AND Wang, Dantong AND Wieland, Franz-Georg AND Banga, Julio R. AND Timmer, Jens AND Villaverde, Alejandro F. AND Sahle, Sven AND Kreutz, Clemens AND Hasenauer, Jan AND Weindl, Daniel},
+    author = {Schmiester, Leonard and Sch{\"a}lte, Yannik and Bergmann, Frank T. and Camba, Tacio and Dudkin, Erika and Egert, Janine and Fr{\"o}hlich, Fabian and Fuhrmann, Lara and Hauber, Adrian L. and Kemmer, Svenja and Lakrisenko, Polina and Loos, Carolin and Merkt, Simon and M{\"u}ller, Wolfgang and Pathirana, Dilan and Raim{\'u}ndez, Elba and Refisch, Lukas and Rosenblatt, Marcus and Stapor, Paul L. and St{\"a}dter, Philipp and Wang, Dantong and Wieland, Franz-Georg and Banga, Julio R. and Timmer, Jens and Villaverde, Alejandro F. and Sahle, Sven and Kreutz, Clemens and Hasenauer, Jan and Weindl, Daniel},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {{PEtab}—Interoperable specification of parameter estimation problems in systems biology},
