+++ doc/JOSS_paper/paper.bib
@@ -80,16 +80,12 @@
 
 @article{SchmiesterSch2021,
     author = {Schmiester, Leonard and Sch{\"a}lte, Yannik and Bergmann, Frank T. and Camba, Tacio and Dudkin, Erika and Egert, Janine and Fr{\"o}hlich, Fabian and Fuhrmann, Lara and Hauber, Adrian L. and Kemmer, Svenja and Lakrisenko, Polina and Loos, Carolin and Merkt, Simon and M{\"u}ller, Wolfgang and Pathirana, Dilan and Raim{\'u}ndez, Elba and Refisch, Lukas and Rosenblatt, Marcus and Stapor, Paul L. and St{\"a}dter, Philipp and Wang, Dantong and Wieland, Franz-Georg and Banga, Julio R. and Timmer, Jens and Villaverde, Alejandro F. and Sahle, Sven and Kreutz, Clemens and Hasenauer, Jan and Weindl, Daniel},
-    journal = {PLOS Computational Biology},
-    publisher = {Public Library of Science},
-    title = {{PEtab}—Interoperable specification of parameter estimation problems in systems biology},
-    year = {2021},
-    month = {01},
-    volume = {17},
-    pages = {1-10},
-    abstract = {Reproducibility and reusability of the results of data-based modeling studies are essential. Yet, there has been—so far—no broadly supported format for the specification of parameter estimation problems in systems biology. Here, we introduce PEtab, a format which facilitates the specification of parameter estimation problems using Systems Biology Markup Language (SBML) models and a set of tab-separated value files describing the observation model and experimental data as well as parameters to be estimated. We already implemented PEtab support into eight well-established model simulation and parameter estimation toolboxes with hundreds of users in total. We provide a Python library for validation and modification of a PEtab problem and currently 20 example parameter estimation problems based on recent studies.},
-    number = {1},
-    doi = {10.1371/journal.pcbi.1008646}
+    journal={PLoS computational biology},
+  volume={17},
+  number={1},
+  pages={e1008646},
+  year={2021},
+  publisher={Public Library of Science San Francisco, CA USA}
 }
 
 @article{BornsteinKea2008,
