+++ doc/JOSS_paper/paper.bib
@@ -79,14 +79,18 @@
 	Year = {2006}}
 
 @article{SchmiesterSch2021,
-    author = {Schmiester, Leonard and Sch{\"a}lte, Yannik and Bergmann, Frank T. and Camba, Tacio and Dudkin, Erika and Egert, Janine and Fr{\"o}hlich, Fabian and Fuhrmann, Lara and Hauber, Adrian L. and Kemmer, Svenja and Lakrisenko, Polina and Loos, Carolin and Merkt, Simon and M{\"u}ller, Wolfgang and Pathirana, Dilan and Raim{\'u}ndez, Elba and Refisch, Lukas and Rosenblatt, Marcus and Stapor, Paul L. and St{\"a}dter, Philipp and Wang, Dantong and Wieland, Franz-Georg and Banga, Julio R. and Timmer, Jens and Villaverde, Alejandro F. and Sahle, Sven and Kreutz, Clemens and Hasenauer, Jan and Weindl, Daniel},
-    journal={PLoS computational biology},
-  volume={17},
-  number={1},
-  pages={e1008646},
-  year={2021},
-  publisher={Public Library of Science San Francisco, CA USA}
-}
+	abstract = {Reproducibility and reusability of the results of data-based modeling studies are essential. Yet, there has been---so far---no broadly supported format for the specification of parameter estimation problems in systems biology. Here, we introduce PEtab, a format which facilitates the specification of parameter estimation problems using Systems Biology Markup Language (SBML) models and a set of tab-separated value files describing the observation model and experimental data as well as parameters to be estimated. We already implemented PEtab support into eight well-established model simulation and parameter estimation toolboxes with hundreds of users in total. We provide a Python library for validation and modification of a PEtab problem and currently 20 example parameter estimation problems based on recent studies.},
+	author = {Schmiester, Leonard AND Sch{\"a}lte, Yannik AND Bergmann, Frank T. AND Camba, Tacio AND Dudkin, Erika AND Egert, Janine AND Fr{\"o}hlich, Fabian AND Fuhrmann, Lara AND Hauber, Adrian L. AND Kemmer, Svenja AND Lakrisenko, Polina AND Loos, Carolin AND Merkt, Simon AND M{\"u}ller, Wolfgang AND Pathirana, Dilan AND Raim{\'u}ndez, Elba AND Refisch, Lukas AND Rosenblatt, Marcus AND Stapor, Paul L. AND St{\"a}dter, Philipp AND Wang, Dantong AND Wieland, Franz-Georg AND Banga, Julio R. AND Timmer, Jens AND Villaverde, Alejandro F. AND Sahle, Sven AND Kreutz, Clemens AND Hasenauer, Jan AND Weindl, Daniel},
+	doi = {10.1371/journal.pcbi.1008646},
+	journal = {PLOS Computational Biology},
+	month = {01},
+	number = {1},
+	pages = {1-10},
+	publisher = {Public Library of Science},
+	title = {{PEtab}---Interoperable specification of parameter estimation problems in systems biology},
+	volume = {17},
+	year = {2021},
+	Bdsk-Url-1 = {https://doi.org/10.1371/journal.pcbi.1008646}}
 
 @article{BornsteinKea2008,
 	An = {PMC2517632},
