### 2021-05-27 #3215 @jvanhoefer
I just noticed, that the reference to the PEtab paper was broken and fixed that. Unfortunately the preview server is (currently) not working for me, which seems to be unrelated to that change.

The rest looks good and I am happy to see the paper accepted :)
