### 2021-05-27 #3215 @arfon
@jvanhoefer – can I just check that you've reviewed the [suggested DOIs from Whedon above](https://github.com/openjournals/joss-reviews/issues/3215#issuecomment-849485218)? If any of them are correct, please add them to your BibTeX file with e.g., `doi = 10.1101/085001` for the first entry.
