### 2021-05-27 9683d12
add DOIs, update AMICI to journal paper
