+++ doc/JOSS_paper/paper.bib
@@ -32,11 +32,19 @@
 	Bdsk-File-1 = {YnBsaXN0MDDSAQIDBFxyZWxhdGl2ZVBhdGhZYWxpYXNEYXRhXxAXUGFwZXJzL0hvb3BzU2FoMjAwNi5wZGZPEQGGAAAAAAGGAAIAAAxNYWNpbnRvc2ggSEQAAAAAAAAAAAAAAAAAAAAAAAAAQkQAAf////8QSG9vcHNTYWgyMDA2LnBkZgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wAAAAAAAAAAAAAAAAABAAMAAAogY3UAAAAAAAAAAAAAAAAABlBhcGVycwACAEUvOlVzZXJzOmphbi5oYXNlbmF1ZXI6c2NpZWJvOk1MUzpCaWJsaW9ncmFwaHk6UGFwZXJzOkhvb3BzU2FoMjAwNi5wZGYAAA4AIgAQAEgAbwBvAHAAcwBTAGEAaAAyADAAMAA2AC4AcABkAGYADwAaAAwATQBhAGMAaQBuAHQAbwBzAGgAIABIAEQAEgBDVXNlcnMvamFuLmhhc2VuYXVlci9zY2llYm8vTUxTL0JpYmxpb2dyYXBoeS9QYXBlcnMvSG9vcHNTYWgyMDA2LnBkZgAAEwABLwAAFQACABT//wAAAAgADQAaACQAPgAAAAAAAAIBAAAAAAAAAAUAAAAAAAAAAAAAAAAAAAHI},
 	Bdsk-Url-1 = {http://dx.doi.org/10.1093/bioinformatics/btl485}}
 	
+	
 @article{FroehlichWei2020,
-  title={AMICI: High-Performance Sensitivity Analysis for Large Ordinary Differential Equation Models},
-  author={Fr{\"o}hlich, Fabian and Weindl, Daniel and Sch{\"a}lte, Yannik and Pathirana, Dilan and Paszkowski, {\L}ukasz and Lines, Glenn Terje and Stapor, Paul and Hasenauer, Jan},
-  journal={arXiv preprint arXiv:2012.09122},
-  year={2020}
+    author = {Fröhlich, Fabian and Weindl, Daniel and Schälte, Yannik and Pathirana, Dilan and Paszkowski, Łukasz and Lines, Glenn Terje and Stapor, Paul and Hasenauer, Jan},
+    title = "{AMICI: high-performance sensitivity analysis for large ordinary differential equation models}",
+    journal = {Bioinformatics},
+    year = {2021},
+    month = {04},
+    abstract = "{Ordinary differential equation models facilitate the understanding of cellular signal transduction and other biological processes. However, for large and comprehensive models, the computational cost of simulating or calibrating can be limiting. AMICI is a modular toolbox implemented in C++/Python/MATLAB that provides efficient simulation and sensitivity analysis routines tailored for scalable, gradient-based parameter estimation and uncertainty quantification.AMICI is published under the permissive BSD-3-Clause license with source code publicly available on https://github.com/AMICI-dev/AMICI. Citeable releases are archived on Zenodo.Supplementary data are available at Bioinformatics online.}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btab227},
+    url = {https://doi.org/10.1093/bioinformatics/btab227},
+    note = {btab227},
[…]
